+++ packages/pluggableWidgets/bar-chart-native/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.2.1] - 2022-6-16
+
+### Fixed
+
+-   Update structure mode.
+-   Update sort description.
+
 ## [2.2.0] - 2022-04-07
 
 ### Added